Optimization.
Once two (or more) disc are properly aligned; at each iteration we
can skip ahead X seconds, where X is the least common multiple of the
sizes of the aligned disc. The given input has all disc have a prime
number as the size, so we can just multiply the sizes.